App Rating Prediction-Python
This project aims to predict the ratings of mobile apps based on various features such as app size, reviews, installs, price, category, and content rating. The goal is to build a predictive model that can identify promising apps that are likely to receive high ratings, which can help the ABC Play Store team in boosting the visibility of certain apps.